Add a sample illustrating client-side SSL/TLS and auth configuration
